Fix package.json and move services
