+++ backend/package.json
@@ -43,7 +43,6 @@
     "@types/multer": "^1.4.7",
     "@types/node-cron": "^3.0.4",
     "@types/nodemailer": "^6.4.5",
-    "@types/pdfkit": "^0.12.7",
     "@types/qrcode": "^1.5.0",
     "@types/uuid": "^8.3.4",
     "@types/xlsx": "^0.0.36",
