+++ backend/src/cronTasks/sendReminderMailTask.ts
@@ -8,7 +8,7 @@ import { User } from '../models/user.model';
 import { Client } from '../models/client.model';
 
 export const sendReminderMailTask = async () => {
-    cron.schedule(`0 0 9-17 * * *`, async () => {
+    cron.schedule(`0 0 ${9 + +process.env.SERVER_TIME_DIFFERENCE}-${17 + +process.env.SERVER_TIME_DIFFERENCE} * * *`, async () => {
         try {
             const orders = await Order.findAll({
                 attributes: { include: [[sequelize.col('Master.name'), 'master'], [sequelize.col('Master.User.email'), 'email'], [sequelize.col('Client.name'), 'client']] },
@@ -16,7 +16,7 @@ export const sendReminderMailTask = async () => {
                 { model: Client, attributes: [] }],
                 where: {
                     date: formatISO(new Date(), { representation: 'date' }),
-                    time: getHours(new Date()) + 1
+                    time: getHours(new Date()) + 1 - +process.env.SERVER_TIME_DIFFERENCE
                 }
             })
             orders.forEach(async (order) => {
