+++ backend/src/services/qrcode.ts
